requirements.txt
torch
transformers
ffmpeg-python
pathlib
